Mover array de peliculas a Controlador
